+++ src/search_inter.h
@@ -71,13 +71,13 @@ typedef double kvz_mvd_cost_func(const encoder_state_t *state,
                                   inter_merge_cand_t merge_cand[MRG_MAX_NUM_CANDS],
                                   int16_t num_cand,
                                   int32_t ref_idx,
-                                  uint32_t *bitcost);
+                                  double *bitcost);
 
 void kvz_search_cu_inter(encoder_state_t * const state,
                          int x, int y, int depth,
                          lcu_t *lcu,
                          double *inter_cost,
-                         uint32_t *inter_bitcost);
+                         double* inter_bitcost);
 
 void kvz_search_cu_smp(encoder_state_t * const state,
                        int x, int y,
@@ -85,7 +85,7 @@ void kvz_search_cu_smp(encoder_state_t * const state,
                        part_mode_t part_mode,
                        lcu_t *lcu,
                        double *inter_cost,
-                       uint32_t *inter_bitcost);
+                       double* inter_bitcost);
 
 
 unsigned kvz_inter_satd_cost(const encoder_state_t* state,
